requirements.txt
aiogram>=3.13.1
tortoise-orm>=0.20.0
aerich>=0.7.2
python-dotenv>=1.0.0
pydantic-settings>=2.5.2
pydantic>=2.10.0
